Rewrite ASTs for analysis phase.
This rewrites the ASTs according to the proposed rules described in #406.

The time to execute unit tests increased from 30 to 32 seconds.
(Actual analysis may be longer)
This hasn't been tested on larger code bases, `'stmts'` may need to be checked
for empty in more places.

In addition to negating conditions, this will split up some `&&` and
move some forms of assignments out of the `if` statement.

Pending tasks:

- if (fn($var = foo())) -> $var = foo(); if (fn($var))
- Make unconditional breaks, continues, throws affect the way
  branched contexts affect merges in analysis phase.
- Investigate false positives this may introduce?
- Add config to disable ast rewriting