Fixes #142 : Automatically inherit types from compatible parent classes
for the `@param` and `@return` annotations.
This also inherits types from interfaces,
or from traits with overridable/abstract methods, etc.

- This is enabled by the `inherit_phpdoc_types` setting, which is true
  by default (Also requires signature compatibility checking to be enabled)
- This may still have edge cases.

If there is already a real type or a phpdoc type on the overriding
method, then Phan won't do anything new. It will keep that type.

- This can be temporarily disabled if there are too many new issues.
- If `return` statements uncover new issues, then fix those,
  or add `@return` if it was a false positive
  (and suppress the phpdoc compatibility warning)
- If parameter types are guessed from parent, causing new issues,
  and those are false positives, adding `@param` on the overriding
  method will help if the guess is incorrect.

This is planned for a subsequent 0.9.3 release.