Warn about undeclared internal functions by default
(even if they aren't available in the binary used to run phan.)

Proposed fix for #1080

The bug that #74 is meant to fix (e.g. for `apache_request_headers`) may affect
only a small fraction of use cases, and can be worked around by adding
a stub function implementation to the parsed files (e.g. in .phan/stubs).

- Inconsistent with what Phan does for classes and methods