Just use plain unique arrays of types, not array sets.
Union types frequently have 0, 1, or 2 types.
Checking if a Type is an array member with
`\in_array($type, $type_list, true)` is faster than doing a hash set
lookup if the number of types are small.
Additionally, many checks are done by a callback.

Runtime for self-analysis decreased with or without `spl_object_id`
available (Comparing this branch and master):

- 5.0 to 4.80 in php 7.2(3-4%), 24.3 to 23.7 seconds analyzing an external project
- 5.7 to 5.2 in php 7.1 (without spl_object_id installed).